IMSS/FinalProject/requirement.txt
python - 3.10.5
keras - 2.9.0
tensorflow-gpu - 2.9.1
gym - 0.24.1
numpy - 1.23.1
pandas - 1.4.3
